backend-fast-api/backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
beanie==1.21.0
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.109.0
flake8==7.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
lazy-model==0.1.0b0
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
motor==3.2.0
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
phonenumbers==9.0.2
platformdirs==4.3.7
pluggy==1.5.0
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.7.0
pydantic-settings==2.3.4
pydantic_core==2.18.1
pyflakes==3.2.0
PyJWT==2.10.1
pymongo==4.5.0
PyMuPDF==1.25.4
PyMySQL==1.1.0
pyparsing==3.2.3
pytest==8.3.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
-e git+https://github.com/ranjeet-h/QP-solver.git@1aa3d76c630007176d7d59466d8114b6c751f2a4#egg=qp_solver_backend&subdirectory=../../..
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.2
starlette==0.35.1
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.23.2
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
